requirements.txt
gradio
huggingface-hub
llama-cpp-python